move Upstream to PaymentPacket
